+++ tic_tac_toe/__main__.py
@@ -1,3 +1,3 @@
-from main import main
+from tic_tac_toe.main import main
 
 main()
